+++ agent/api/task/task_linux_test.go
@@ -492,7 +492,7 @@ func TestPlatformHostConfigOverrideErrorPath(t *testing.T) {
 	assert.Empty(t, dockerHostConfig)
 }
 
-func TestDockerHostConfigRawConfigMerging(t *testing.T) {
+func skipTestDockerHostConfigRawConfigMerging(t *testing.T) {
 	// Use a struct that will marshal to the actual message we expect; not
 	// dockercontainer.HostConfig which will include a lot of zero values.
 	rawHostConfigInput := struct {
