+++ agent/app/agent_capability_unix_test.go
@@ -853,16 +853,6 @@ func TestFirelensConfigCapabilitiesUnix(t *testing.T) {
 	assert.Contains(t, capabilities, &ecs.Attribute{Name: aws.String(attributePrefix + capabilityFirelensConfigS3)})
 }
 
-func TestAppendGMSACapabilities(t *testing.T) {
-	var inputCapabilities []*ecs.Attribute
-
-	agent := &ecsAgent{}
-
-	capabilities := agent.appendGMSACapabilities(inputCapabilities)
-	assert.Equal(t, len(inputCapabilities), len(capabilities))
-	assert.EqualValues(t, capabilities, inputCapabilities)
-}
-
 func TestAppendFSxWindowsFileServerCapabilities(t *testing.T) {
 	var inputCapabilities []*ecs.Attribute
